move asDict methods to test suite
