+++ python/pyspark/sql/streaming/listener.py
@@ -22,15 +22,6 @@ from abc import ABC, abstractmethod
 from py4j.java_gateway import JavaObject
 
 from pyspark.sql import Row
-from pyspark.sql.types import (
-    ArrayType,
-    StructType,
-    StructField,
-    StringType,
-    IntegerType,
-    FloatType,
-    MapType,
-)
 from pyspark import cloudpickle
 
 __all__ = ["StreamingQueryListener"]
@@ -206,15 +197,6 @@ class QueryStartedEvent:
             timestamp=j["timestamp"],
         )
 
-    def asDict(self) -> Dict[str, Any]:
-        def conv(obj: Any) -> Any:
-            if isinstance(obj, uuid.UUID):
-                return str(obj)
-            else:
-                return obj
-
-        return {k[1:]: conv(v) for k, v in self.__dict__.items()}
-
     @property
     def id(self) -> uuid.UUID:
         """
@@ -275,9 +257,6 @@ class QueryProgressEvent:
         """
         return self._progress
 
-    def asDict(self) -> Dict[str, Any]:
-        return {"progress": self.progress.asDict()}
-
 
 class QueryIdleEvent:
     """
@@ -306,15 +285,6 @@ class QueryIdleEvent:
     @classmethod
     def fromJson(cls, j: Dict[str, Any]) -> "QueryIdleEvent":
         return cls(id=uuid.UUID(j["id"]), runId=uuid.UUID(j["runId"]), timestamp=j["timestamp"])
-
-    def asDict(self) -> Dict[str, Any]:
-        def conv(obj: Any) -> Any:
-            if isinstance(obj, uuid.UUID):
-                return str(obj)
-            else:
-                return obj
-
-        return {k[1:]: conv(v) for k, v in self.__dict__.items()}
 
     @property
     def id(self) -> uuid.UUID:
@@ -382,15 +352,6 @@ class QueryTerminatedEvent:
             exception=j["exception"],
             errorClassOnException=j["errorClassOnException"],
         )
-
-    def asDict(self) -> Dict[str, Any]:
-        def conv(obj: Any) -> Any:
-            if isinstance(obj, uuid.UUID):
-                return str(obj)
-            else:
-                return obj
-
-        return {k[1:]: conv(v) for k, v in self.__dict__.items()}
 
     @property
     def id(self) -> uuid.UUID:
@@ -534,25 +495,6 @@ class StreamingQueryProgress:
             if "observedMetrics" in j
             else {},
         )
-
-    def asDict(self) -> Dict[str, Any]:
-        def conv(obj: Any) -> Any:
-            if isinstance(obj, uuid.UUID):
-                return str(obj)
-            elif isinstance(obj, (SourceProgress, SinkProgress, StateOperatorProgress)):
-                return obj.asDict()
-            elif isinstance(obj, Row):
-                return json.dumps(obj.asDict())  # Assume no nested row in observed metrics
-            elif isinstance(obj, list):
-                return [conv(o) for o in obj]
-            elif isinstance(obj, dict):
-                return dict((k, conv(v)) for k, v in obj.items())
-            else:
-                return obj
-
-        return {
-            k[1:]: conv(v) for k, v in self.__dict__.items() if k not in ["_jprogress", "_jdict"]
-        }
 
     @property
     def id(self) -> uuid.UUID:
@@ -776,9 +718,6 @@ class StateOperatorProgress:
             customMetrics=dict(j["customMetrics"]) if "customMetrics" in j else {},
         )
 
-    def asDict(self) -> Dict[str, Any]:
-        return {k[1:]: v for k, v in self.__dict__.items() if k not in ["_jprogress", "_jdict"]}
-
     @property
     def operatorName(self) -> str:
         return self._operatorName
@@ -914,9 +853,6 @@ class SourceProgress:
             metrics=dict(j["metrics"]) if "metrics" in j else {},
         )
 
-    def asDict(self) -> Dict[str, Any]:
-        return {k[1:]: v for k, v in self.__dict__.items() if k not in ["_jprogress", "_jdict"]}
-
     @property
     def description(self) -> str:
         """
@@ -1027,9 +963,6 @@ class SinkProgress:
             numOutputRows=jprogress.numOutputRows(),
             metrics=dict(jprogress.metrics()),
         )
-
-    def asDict(self) -> Dict[str, Any]:
-        return {k[1:]: v for k, v in self.__dict__.items() if k not in ["_jprogress", "_jdict"]}
 
     @classmethod
     def fromJson(cls, j: Dict[str, Any]) -> "SinkProgress":
